utility/requirements.txt
tqdm~=4.64.1
Pillow~=9.3.0